- Silent warning about comparison between signed and unsigned integer
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@10777 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/sdk/findreplacedlg.cpp
@@ -183,7 +183,7 @@ FindReplaceDlg::FindReplaceDlg(wxWindow* parent, const wxString& initial, bool h
     wxRadioBox* rbScope = XRCCTRL(*this, "rbScope2", wxRadioBox);
     EditorManager* edMgr = Manager::Get()->GetEditorManager();
     bool filesOpen = false;
-    for (size_t i = 0; i < edMgr->GetEditorsCount(); ++i)
+    for (int i = 0; i < edMgr->GetEditorsCount(); ++i)
     {
         if (edMgr->GetBuiltinEditor(i))
         {
